refactor: 불명확한 네이밍 수정
