+++ src/main/java/controller/LottoController.java
@@ -3,8 +3,8 @@ package controller;
 import domain.dto.WinningNumbersDto;
 import domain.lotto.LottoGame;
 import domain.lotto.WinningNumbers;
-import domain.lotto.number.LottoNumberMaker;
-import domain.lotto.number.RandomNumberGenerator;
+import domain.lotto.makingnumbers.LottoNumberMaker;
+import domain.lotto.makingnumbers.RandomNumberGenerator;
 import domain.lotto.result.LottoResults;
 import domain.lotto.ticket.LottoTicket;
 import domain.lotto.ticket.LottoTickets;
