+++ src/main/java/domain/lotto/makingnumbers/RandomNumberGenerator.java
@@ -1,6 +1,6 @@
-package domain.lotto.number;
+package domain.lotto.makingnumbers;
 
-import domain.lotto.LottoMetaData;
+import domain.lotto.LottoConstantData;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -12,11 +12,11 @@ public class RandomNumberGenerator implements NumberGeneratable {
     public List<Integer> generate() {
         List<Integer> list = new ArrayList<>();
 
-        for (int i = LottoMetaData.MIN_LOTTO_NUMBER; i <= LottoMetaData.MAX_LOTTO_NUMBER; i++) {
+        for (int i = LottoConstantData.MIN_LOTTO_NUMBER; i <= LottoConstantData.MAX_LOTTO_NUMBER; i++) {
             list.add(i);
         }
 
         Collections.shuffle(list);
-        return list.subList(0, LottoMetaData.LOTTO_NUMBER_SIZE);
+        return list.subList(0, LottoConstantData.LOTTO_NUMBER_SIZE);
     }
 }
