+++ src/main/java/domain/lotto/result/LottoResults.java
@@ -1,8 +1,7 @@
 package domain.lotto.result;
 
-import domain.lotto.LottoMetaData;
+import domain.lotto.LottoConstantData;
 
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -28,7 +27,7 @@ public class LottoResults {
         Integer ticketNumber = lottoResultCount.keySet().stream()
                 .mapToInt(lottoResultCount::get)
                 .sum();
-        Integer budget = ticketNumber * LottoMetaData.LOTTO_TICKET_PRICE;
+        Integer budget = ticketNumber * LottoConstantData.LOTTO_TICKET_PRICE;
 
         Integer profit = lottoResultCount.keySet().stream()
                 .mapToInt((key) -> lottoResultCount.get(key) * key.getPrize())
